Type/Class cleanup: Steam.AppLists.cs

+++ addons/godotsteam_csharpbindings/Steam.AppLists.Signals.cs
@@ -5,9 +5,9 @@ namespace GodotSteam;
 
 public static partial class Steam
 {
-    public delegate void AppInstalledEventHandler(uint appId, long installFolderIndex);
+    public delegate void AppInstalledEventHandler(uint appId, uint installFolderIndex);
     private static event AppInstalledEventHandler AppInstalledEvent;
-    static Action<uint, long> _appInstalledAction = (appId, installFolderIndex) =>
+    static Action<uint, uint> _appInstalledAction = (appId, installFolderIndex) =>
     {
         AppInstalledEvent?.Invoke(appId, installFolderIndex);
     };
@@ -31,9 +31,9 @@ public static partial class Steam
         }
     }
     
-    public delegate void AppUninstalledEventHandler(uint appId, long installFolderIndex);
+    public delegate void AppUninstalledEventHandler(uint appId, uint installFolderIndex);
     private static event AppUninstalledEventHandler AppUninstalledEvent;
-    static Action<uint, long> _appUninstalledAction = (appId, installFolderIndex) =>
+    static Action<uint, uint> _appUninstalledAction = (appId, installFolderIndex) =>
     {
         AppUninstalledEvent?.Invoke(appId, installFolderIndex);
     };
